feat: Extract view interface for the web ui later
