+++ src/main/java/com/woowacourse/lotto/Main.java
@@ -3,23 +3,28 @@ package com.woowacourse.lotto;
 import com.woowacourse.lotto.domain.*;
 import com.woowacourse.lotto.service.LottoService;
 import com.woowacourse.lotto.view.InputView;
+import com.woowacourse.lotto.view.InputViewConsole;
 import com.woowacourse.lotto.view.OutputView;
+import com.woowacourse.lotto.view.OutputViewConsole;
 
 public class Main {
     public static void main(String[] args) {
-        PurchaseMoney purchaseMoney = InputView.getMoneyFromUser();
-        ManualNumber manualNumber = InputView.getManualNumberFromUser(purchaseMoney);
+        InputView inputViewConsole = new InputViewConsole();
+        OutputView outputViewConsole = new OutputViewConsole();
 
-        LottoBuyList manualBuyList = InputView.getManualLottoFromUser(manualNumber);
+        PurchaseMoney purchaseMoney = inputViewConsole.getMoneyFromUser();
+        ManualNumber manualNumber = inputViewConsole.getManualNumberFromUser(purchaseMoney);
+
+        LottoBuyList manualBuyList = inputViewConsole.getManualLottoFromUser(manualNumber);
         LottoBuyList autoBuyList = LottoService.getAutoLottoBuyList(purchaseMoney, manualNumber);
         LottoBuyList totalBuyList = LottoService.joinBuyList(manualBuyList, autoBuyList);
 
-        OutputView.printLottoBuyList(totalBuyList);
+        outputViewConsole.printLottoBuyList(totalBuyList);
 
-        Lotto lastWeekLotto = InputView.getLastWeekLottoFromUser();
-        LottoNumber bonusNumber = InputView.getBonusNumberFromUser(lastWeekLotto);
+        Lotto lastWeekLotto = inputViewConsole.getLastWeekLottoFromUser();
+        LottoNumber bonusNumber = inputViewConsole.getBonusNumberFromUser(lastWeekLotto);
         LottoResult lottoResult = LottoService.getLottoResult(totalBuyList, new WinningLotto(lastWeekLotto, bonusNumber));
 
-        OutputView.printLottoResult(lottoResult, purchaseMoney);
+        outputViewConsole.printLottoResult(lottoResult, purchaseMoney);
     }
 }
